Added support for ruby 2.3 version
Quick fix in init.d script to make it work with 2.3 ruby version. The permanent fix would be to make it as erb template and parameterize the ruby version.  The other way around could be checking the current ruby version and if that is not included in the list then append as well.